ingress/requirements.txt
google-cloud-datastore==1.7.1
google-cloud-storage==1.13.0
google-cloud-pubsub==0.39.1
google-cloud-logging==1.8.0
google-cloud-error_reporting==0.30.0
pandas 
rethinkdb
ta
matplotlib
numpy